+++ src/components/App.js
@@ -1,10 +1,10 @@
 import GlobalStyles from "../styles/GlobalStyles";
 import {
-    BrowserRouter,
-    Routes,
-    Route,
-    useLocation,
-    Navigate,
+  BrowserRouter,
+  Routes,
+  Route,
+  useLocation,
+  Navigate,
 } from "react-router-dom";
 import SignUp from "./Signup/Signup.js";
 import SignIn from "./Signin/Signin";
@@ -14,44 +14,36 @@ import Header from "./Header";
 import Home from "./home/HomePage";
 
 function App() {
-    const location = useLocation();
-    const localStorageUser = JSON.parse(localStorage.getItem("userLinkr"));
-    const [loggedUser, setLoggedUser] = useState(localStorageUser);
+  const location = useLocation();
+  const localStorageUser = JSON.parse(localStorage.getItem("userLinkr"));
+  const [loggedUser, setLoggedUser] = useState(localStorageUser);
 
-    const [users, setUsers] = useState();
-    return (
-        <>
-            <UserContext.Provider value={{ users, setUsers, setLoggedUser }}>
-                <GlobalStyles />
-                <Routes location={location} key={location.pathname}>
-                    <Route
-                        path='/home'
-                        element={
-                            loggedUser ? (
-                                <>
-                                    <Header />
-                                    <Home />
-                                </>
-                            ) : (
-                                <Navigate replace to={"/"} />
-                            )
-                        }
-                    />
-                    <Route path='/sign-up/' element={<SignUp />} />
-                    <Route
-                        path='/'
-                        element={
-                            loggedUser ? (
-                                <Navigate replace to={"/home"} />
-                            ) : (
-                                <SignIn />
-                            )
-                        }
-                    />
-                </Routes>
-            </UserContext.Provider>
-        </>
-    );
+  const [users, setUsers] = useState();
+  return (
+    <>
+      <UserContext.Provider value={{ users, setUsers, setLoggedUser }}>
+        <GlobalStyles />
+        <Routes location={location} key={location.pathname}>
+          <Route
+            path="/home"
+            element={
+              <>
+                <Header />
+                <Home />
+              </>
+            }
+          />
+          <Route path="/sign-up/" element={<SignUp />} />
+          <Route
+            path="/"
+            element={
+              loggedUser ? <Navigate replace to={"/home"} /> : <SignIn />
+            }
+          />
+        </Routes>
+      </UserContext.Provider>
+    </>
+  );
 }
 
 export default App;
